chore: rename global style file
