+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from 'next/app'
 import React from 'react'
 
-import GlobalStyle from '../styles/globalStyles'
+import GlobalStyle from '../styles/global'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
